requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
openai==1.3.7
python-dotenv==1.0.0
requests==2.31.0
yfinance==0.2.28
pandas==2.1.4
numpy==1.24.3
asyncio==3.4.3
aiohttp==3.9.1
pydantic==2.5.2
typing-extensions==4.8.0 